Attempts to speed up generation of latlon grid
Potentially unworkable bottleneck without resorting to Cython - might roll back

I've tried a few things to get the lat-lon transform running at the
same speed as the rest of the processing chain, but it looks like
osr.CoordinateTrasform.TrasformPoint() is the bottleneck.

It
-cannot be vectorised (not a Numpy-based function)
-cannot be parallelized using Loky or Multiprocess (can't pickle a Swig
object, which is what CoordinateTransform is....but maybe you can create a
new one in each thread?)
-cannot be multithreaded, due to (I think) Swig not releasing the GIL

(if you are reading this and it's all voodoo, no worries. If you are
curious, start here: https://realpython.com/python-gil/)